HEEDLS-581 Fix circular dependency and unit tests
