+++ DigitalLearningSolutions.Data.Tests/Services/LoginServiceTests.cs
@@ -3,7 +3,6 @@
     using System;
     using System.Collections.Generic;
     using System.Linq;
-    using DigitalLearningSolutions.Data.DataServices.UserDataService;
     using DigitalLearningSolutions.Data.Enums;
     using DigitalLearningSolutions.Data.Models.User;
     using DigitalLearningSolutions.Data.Services;
@@ -17,19 +16,17 @@
     {
         private const string Username = "Username";
         private const string Password = "Password";
-        private ICryptoService cryptoService = null!;
         private LoginService loginService = null!;
-        private IUserDataService userDataService = null!;
         private IUserService userService = null!;
+        private IUserVerificationService userVerificationService = null!;
 
         [SetUp]
         public void Setup()
         {
-            userDataService = A.Fake<IUserDataService>(x => x.Strict());
-            cryptoService = A.Fake<ICryptoService>(x => x.Strict());
+            userVerificationService = A.Fake<IUserVerificationService>(x => x.Strict());
             userService = A.Fake<IUserService>(x => x.Strict());
 
-            loginService = new LoginService(userService, userDataService, cryptoService);
+            loginService = new LoginService(userService, userVerificationService);
         }
 
         [Test]
@@ -58,8 +55,8 @@
             var delegateUser = UserTestHelper.GetDefaultDelegateUser();
             A.CallTo(() => userService.GetUsersByUsername(Username))
                 .Returns((new List<AdminUser>(), new List<DelegateUser> { delegateUser }));
-            A.CallTo(() => cryptoService.VerifyHashedPassword(delegateUser.Password!, Password))
-                .Returns(false);
+            A.CallTo(() => userVerificationService.VerifyUsers(Password, A<List<AdminUser>>._, A<List<DelegateUser>>._))
+                .Returns(new UserAccountSet(new List<AdminUser>(), new List<DelegateUser>()));
 
             // When
             var result = loginService.AttemptLogin(Username, Password);
@@ -80,8 +77,8 @@
             var adminUser = UserTestHelper.GetDefaultAdminUser();
             A.CallTo(() => userService.GetUsersByUsername(Username))
                 .Returns((new List<AdminUser> { adminUser }, new List<DelegateUser>()));
-            A.CallTo(() => cryptoService.VerifyHashedPassword(adminUser.Password!, Password))
-                .Returns(false);
+            A.CallTo(() => userVerificationService.VerifyUsers(Password, A<List<AdminUser>>._, A<List<DelegateUser>>._))
+                .Returns(new UserAccountSet(new List<AdminUser>(), new List<DelegateUser>()));
             A.CallTo(() => userService.IncrementFailedLoginCount(adminUser)).DoesNothing();
 
             // When
@@ -104,8 +101,8 @@
             var adminUser = UserTestHelper.GetDefaultAdminUser(failedLoginCount: 4);
             A.CallTo(() => userService.GetUsersByUsername(Username))
                 .Returns((new List<AdminUser> { adminUser }, new List<DelegateUser>()));
-            A.CallTo(() => cryptoService.VerifyHashedPassword(adminUser.Password!, Password))
-                .Returns(false);
+            A.CallTo(() => userVerificationService.VerifyUsers(Password, A<List<AdminUser>>._, A<List<DelegateUser>>._))
+                .Returns(new UserAccountSet(new List<AdminUser>(), new List<DelegateUser>()));
             A.CallTo(() => userService.IncrementFailedLoginCount(adminUser)).DoesNothing();
 
             // When
@@ -128,8 +125,8 @@
             var adminUser = UserTestHelper.GetDefaultAdminUser(failedLoginCount: 6);
             A.CallTo(() => userService.GetUsersByUsername(Username))
                 .Returns((new List<AdminUser> { adminUser }, new List<DelegateUser>()));
-            A.CallTo(() => cryptoService.VerifyHashedPassword(adminUser.Password!, Password))
-                .Returns(false);
+            A.CallTo(() => userVerificationService.VerifyUsers(Password, A<List<AdminUser>>._, A<List<DelegateUser>>._))
+                .Returns(new UserAccountSet(new List<AdminUser>(), new List<DelegateUser>()));
             A.CallTo(() => userService.IncrementFailedLoginCount(adminUser)).DoesNothing();
 
             // When
@@ -152,8 +149,8 @@
             var delegateUser = UserTestHelper.GetDefaultDelegateUser(approved: false);
             A.CallTo(() => userService.GetUsersByUsername(Username))
                 .Returns((new List<AdminUser>(), new List<DelegateUser> { delegateUser }));
-            A.CallTo(() => cryptoService.VerifyHashedPassword(delegateUser.Password!, Password))
-                .Returns(true);
+            A.CallTo(() => userVerificationService.VerifyUsers(Password, A<List<AdminUser>>._, A<List<DelegateUser>>._))
+                .Returns(new UserAccountSet(new List<AdminUser>(), new List<DelegateUser> { delegateUser }));
 
             // When
             var result = loginService.AttemptLogin(Username, Password);
@@ -176,8 +173,8 @@
             var adminUsers = new List<AdminUser> { adminUser, secondAdminUser };
             A.CallTo(() => userService.GetUsersByUsername(Username))
                 .Returns((adminUsers, new List<DelegateUser>()));
-            A.CallTo(() => cryptoService.VerifyHashedPassword(A<string?>._, Password))
-                .Returns(true);
+            A.CallTo(() => userVerificationService.VerifyUsers(Password, A<List<AdminUser>>._, A<List<DelegateUser>>._))
+                .Returns(new UserAccountSet(adminUsers, new List<DelegateUser>()));
 
             // Then
             Assert.Throws<InvalidOperationException>(() => loginService.AttemptLogin(Username, Password));
@@ -191,8 +188,8 @@
             var delegateUser = UserTestHelper.GetDefaultDelegateUser();
             A.CallTo(() => userService.GetUsersByUsername(Username))
                 .Returns((new List<AdminUser> { adminUser }, new List<DelegateUser> { delegateUser }));
-            A.CallTo(() => cryptoService.VerifyHashedPassword(A<string?>._, Password))
-                .Returns(true);
+            A.CallTo(() => userVerificationService.VerifyUsers(Password, A<List<AdminUser>>._, A<List<DelegateUser>>._))
+                .Returns(new UserAccountSet(new List<AdminUser> { adminUser }, new List<DelegateUser> { delegateUser }));
 
             // Then
             var exception = Assert.Throws<Exception>(() => loginService.AttemptLogin(Username, Password));
@@ -207,11 +204,15 @@
             var adminUser = UserTestHelper.GetDefaultAdminUser(emailAddress: "email@test.com");
             A.CallTo(() => userService.GetUsersByUsername(Username))
                 .Returns((new List<AdminUser> { adminUser }, new List<DelegateUser>()));
-            A.CallTo(() => cryptoService.VerifyHashedPassword(A<string?>._, Password))
-                .Returns(true);
+            A.CallTo(() => userVerificationService.VerifyUsers(Password, A<List<AdminUser>>._, A<List<DelegateUser>>._))
+                .Returns(new UserAccountSet(new List<AdminUser> { adminUser }, new List<DelegateUser>()));
             A.CallTo(() => userService.ResetFailedLoginCount(adminUser)).DoesNothing();
-            A.CallTo(() => userDataService.GetDelegateUsersByEmailAddress(adminUser.EmailAddress!))
-                .Returns(new List<DelegateUser>());
+            A.CallTo(
+                () => userVerificationService.GetVerifiedDelegateUsersAssociatedWithAdminUser(
+                    adminUser,
+                    Password
+                )
+            ).Returns(new List<DelegateUser>());
             A.CallTo(() => userService.GetUsersWithActiveCentres(adminUser, A<List<DelegateUser>>._))
                 .Returns((adminUser, new List<DelegateUser>()));
             A.CallTo(() => userService.GetUserCentres(adminUser, A<List<DelegateUser>>._)).Returns(
@@ -224,7 +225,12 @@
             // Then
             using (new AssertionScope())
             {
-                A.CallTo(() => userDataService.GetAdminUserByEmailAddress(A<string>._)).MustNotHaveHappened();
+                A.CallTo(
+                    () => userVerificationService.GetVerifiedAdminUserAssociatedWithDelegateUser(
+                        A<DelegateUser?>._,
+                        A<string>._
+                    )
+                ).MustNotHaveHappened();
                 result.LoginAttemptResult.Should().Be(LoginAttemptResult.LogIntoSingleCentre);
                 result.LogInAdmin.Should().Be(adminUser);
                 result.LogInDelegates.Should().BeEmpty();
@@ -240,14 +246,25 @@
             var delegateUser = UserTestHelper.GetDefaultDelegateUser();
             A.CallTo(() => userService.GetUsersByUsername(Username))
                 .Returns((new List<AdminUser>(), new List<DelegateUser> { delegateUser }));
-            A.CallTo(() => cryptoService.VerifyHashedPassword(A<string?>._, Password))
-                .Returns(true);
-            A.CallTo(() => userDataService.GetAdminUserByEmailAddress(delegateUser.EmailAddress!))
-                .Returns(linkedAdminUser);
+            A.CallTo(() => userVerificationService.VerifyUsers(Password, A<List<AdminUser>>._, A<List<DelegateUser>>._))
+                .Returns(new UserAccountSet(new List<AdminUser>(), new List<DelegateUser> { delegateUser }));
+            A.CallTo(
+                () => userVerificationService.GetVerifiedAdminUserAssociatedWithDelegateUser(
+                    delegateUser,
+                    Password
+                )
+            ).Returns(linkedAdminUser);
+            A.CallTo(
+                () => userVerificationService.GetVerifiedDelegateUsersAssociatedWithAdminUser(
+                    null,
+                    Password
+                )
+            ).Returns(new List<DelegateUser>());
             A.CallTo(() => userService.GetUsersWithActiveCentres(linkedAdminUser, A<List<DelegateUser>>._))
                 .Returns((linkedAdminUser, new List<DelegateUser> { delegateUser }));
             A.CallTo(() => userService.GetUserCentres(linkedAdminUser, A<List<DelegateUser>>._)).Returns(
-                new List<CentreUserDetails> { new CentreUserDetails(linkedAdminUser.CentreId, linkedAdminUser.CentreName, true) }
+                new List<CentreUserDetails>
+                    { new CentreUserDetails(linkedAdminUser.CentreId, linkedAdminUser.CentreName, true) }
             );
 
             // When
@@ -261,311 +278,5 @@
                 result.LogInDelegates.Single().Should().Be(delegateUser);
             }
         }
-
-        #region old tests
-
-        [Test]
-        public void VerifyUsers_Returns_verified_admin_user()
-        {
-            // Given
-            A.CallTo(() => cryptoService.VerifyHashedPassword(A<string>._, A<string>._)).Returns(false);
-            A.CallTo(() => cryptoService.VerifyHashedPassword("Automatically Verified", A<string>._)).Returns(true);
-            var adminUser = UserTestHelper.GetDefaultAdminUser(password: "Automatically Verified");
-            var adminUsers = new List<AdminUser> { adminUser };
-
-            //When
-            var (verifiedAdminUsers, _) = loginService.VerifyUsers(
-                "password",
-                adminUsers,
-                new List<DelegateUser>()
-            );
-
-            // Then
-            Assert.AreEqual(adminUser, verifiedAdminUsers.First());
-        }
-
-        [Test]
-        public void VerifyUsers_Does_not_return_unverified_admin_user()
-        {
-            // Given
-            A.CallTo(() => cryptoService.VerifyHashedPassword(A<string>._, A<string>._)).Returns(false);
-
-            var adminUser = UserTestHelper.GetDefaultAdminUser();
-            var adminUsers = new List<AdminUser> { adminUser };
-
-            //When
-            var (verifiedAdminUsers, _) = loginService.VerifyUsers(
-                "password",
-                adminUsers,
-                new List<DelegateUser>()
-            );
-
-            // Then
-            Assert.IsEmpty(verifiedAdminUsers);
-        }
-
-        [Test]
-        public void VerifyUsers__Returns_verified_delegate_users()
-        {
-            // Given
-            A.CallTo(() => cryptoService.VerifyHashedPassword(A<string>._, A<string>._)).Returns(false);
-            A.CallTo(() => cryptoService.VerifyHashedPassword("Automatically Verified", A<string>._)).Returns(true);
-
-            var firstDelegateUser = UserTestHelper.GetDefaultDelegateUser(password: "Automatically Verified", id: 1);
-            var secondDelegateUser = UserTestHelper.GetDefaultDelegateUser(password: "Fails Verification", id: 2);
-            var thirdDelegateUser = UserTestHelper.GetDefaultDelegateUser(password: "Automatically Verified", id: 3);
-
-            var delegateUsers = new List<DelegateUser>
-            {
-                firstDelegateUser,
-                secondDelegateUser,
-                thirdDelegateUser
-            };
-
-            var adminUsers = new List<AdminUser> { UserTestHelper.GetDefaultAdminUser() };
-
-            //When
-            var (_, verifiedDelegateUsers) = loginService.VerifyUsers(
-                "password",
-                adminUsers,
-                delegateUsers
-            );
-
-            // Then
-            Assert.Contains(firstDelegateUser, verifiedDelegateUsers);
-            Assert.Contains(thirdDelegateUser, verifiedDelegateUsers);
-        }
-
-        [Test]
-        public void VerifyUsers_Filters_out_unverified_delegate_users()
-        {
-            // Given
-            A.CallTo(() => cryptoService.VerifyHashedPassword(A<string>._, A<string>._)).Returns(false);
-            A.CallTo(() => cryptoService.VerifyHashedPassword("Automatically Verified", A<string>._)).Returns(true);
-
-            var firstDelegateUser = UserTestHelper.GetDefaultDelegateUser(password: "Automatically Verified", id: 1);
-            var secondDelegateUser = UserTestHelper.GetDefaultDelegateUser(password: "Fails Verification", id: 2);
-            var thirdDelegateUser = UserTestHelper.GetDefaultDelegateUser(password: "Automatically Verified", id: 3);
-
-            var delegateUsers = new List<DelegateUser>
-            {
-                firstDelegateUser,
-                secondDelegateUser,
-                thirdDelegateUser
-            };
-
-            var adminUsers = new List<AdminUser> { UserTestHelper.GetDefaultAdminUser() };
-
-            //When
-            var (_, verifiedDelegateUsers) = loginService.VerifyUsers(
-                "password",
-                adminUsers,
-                delegateUsers
-            );
-
-            // Then
-            Assert.IsFalse(verifiedDelegateUsers.Contains(secondDelegateUser));
-        }
-
-        [Test]
-        public void VerifyUsers_Returns_no_delegates_if_all_unverified()
-        {
-            // Given
-            A.CallTo(() => cryptoService.VerifyHashedPassword(A<string>._, A<string>._)).Returns(false);
-
-            var firstDelegateUser = UserTestHelper.GetDefaultDelegateUser(1);
-            var secondDelegateUser = UserTestHelper.GetDefaultDelegateUser();
-            var thirdDelegateUser = UserTestHelper.GetDefaultDelegateUser(3);
-
-            var delegateUsers = new List<DelegateUser>
-            {
-                firstDelegateUser,
-                secondDelegateUser,
-                thirdDelegateUser
-            };
-
-            var adminUsers = new List<AdminUser> { UserTestHelper.GetDefaultAdminUser() };
-
-            //When
-            var (_, verifiedDelegateUsers) = loginService.VerifyUsers(
-                "password",
-                adminUsers,
-                delegateUsers
-            );
-
-            // Then
-            Assert.IsEmpty(verifiedDelegateUsers);
-        }
-
-        [Test]
-        public void VerifyUsers_Returns_no_delegate_users_when_delegate_stored_password_is_empty()
-        {
-            // Given
-            A.CallTo(() => cryptoService.VerifyHashedPassword(A<string>._, A<string>._)).Returns(false);
-            var delegateUserWithoutPassword = UserTestHelper.GetDefaultDelegateUser(password: string.Empty);
-            var adminUsers = new List<AdminUser> { UserTestHelper.GetDefaultAdminUser() };
-
-            //When
-            var (_, returnedDelegateList) = loginService.VerifyUsers(
-                "password",
-                adminUsers,
-                new List<DelegateUser> { delegateUserWithoutPassword }
-            );
-
-            // Then
-            Assert.IsEmpty(returnedDelegateList);
-        }
-
-        [Test]
-        public void
-            GetVerifiedAdminUserAssociatedWithDelegateUsers_Returns_nothing_when_delegate_email_and_alias_is_empty()
-        {
-            // Given
-            var delegateUser = UserTestHelper.GetDefaultDelegateUser(emailAddress: null, aliasId: null);
-            var delegateUsers = new List<DelegateUser> { delegateUser };
-
-            // When
-            var returnedAdminUser = loginService.GetVerifiedAdminUserAssociatedWithDelegateUsers(
-                delegateUsers,
-                "password"
-            );
-
-            // Then
-            Assert.IsNull(returnedAdminUser);
-        }
-
-        [Test]
-        public void
-            GetVerifiedAdminUserAssociatedWithDelegateUsers_Returns_nothing_when_no_admin_account_is_associated_with_delegate()
-        {
-            // Given
-            var delegateUser = UserTestHelper.GetDefaultDelegateUser();
-            var delegateUsers = new List<DelegateUser> { delegateUser };
-            A.CallTo(() => userDataService.GetAdminUserByEmailAddress(A<string>._)).Returns(null);
-
-            // When
-            var returnedAdminUser = loginService.GetVerifiedAdminUserAssociatedWithDelegateUsers(
-                delegateUsers,
-                "password"
-            );
-
-            // Then
-            Assert.IsNull(returnedAdminUser);
-        }
-
-        [Test]
-        public void
-            GetVerifiedAdminUserAssociatedWithDelegateUsers_Returns_nothing_when_admin_account_associated_only_by_email_is_at_different_centre()
-        {
-            // Given
-            var delegateUser = UserTestHelper.GetDefaultDelegateUser(centreId: 2);
-            var delegateUsers = new List<DelegateUser> { delegateUser };
-            var associatedAdminUser = UserTestHelper.GetDefaultAdminUser(centreId: 5);
-            A.CallTo(() => userDataService.GetAdminUserByEmailAddress(A<string>._)).Returns(associatedAdminUser);
-            A.CallTo(() => cryptoService.VerifyHashedPassword(A<string>._, A<string>._)).Returns(true);
-
-            // When
-            var returnedAdminUser = loginService.GetVerifiedAdminUserAssociatedWithDelegateUsers(
-                delegateUsers,
-                "password"
-            );
-
-            // Then
-            Assert.IsNull(returnedAdminUser);
-        }
-
-        [Test]
-        public void
-            GetVerifiedAdminUserAssociatedWithDelegateUsers_Returns_nothing_when_admin_account_associated_only_by_alias_is_at_different_centre()
-        {
-            // Given
-            var delegateUser = UserTestHelper.GetDefaultDelegateUser(centreId: 2, emailAddress: null, aliasId: "Test");
-            var delegateUsers = new List<DelegateUser> { delegateUser };
-            var associatedAdminUser = UserTestHelper.GetDefaultAdminUser(centreId: 5);
-            A.CallTo(() => userDataService.GetAdminUserByEmailAddress(A<string>._)).Returns(associatedAdminUser);
-            A.CallTo(() => cryptoService.VerifyHashedPassword(A<string>._, A<string>._)).Returns(true);
-
-            // When
-            var returnedAdminUser = loginService.GetVerifiedAdminUserAssociatedWithDelegateUsers(
-                delegateUsers,
-                "password"
-            );
-
-            // Then
-            Assert.IsNull(returnedAdminUser);
-        }
-
-        [Test]
-        public void
-            GetVerifiedAdminUserAssociatedWithDelegateUsers_Returns_verified_admin_account_associated_with_delegate_by_email()
-        {
-            // Given
-            var delegateUser = UserTestHelper.GetDefaultDelegateUser();
-            var delegateUsers = new List<DelegateUser> { delegateUser };
-            var associatedAdminUser = UserTestHelper.GetDefaultAdminUser();
-            A.CallTo(() => userDataService.GetAdminUserByEmailAddress(A<string>._)).Returns(associatedAdminUser);
-            A.CallTo(() => cryptoService.VerifyHashedPassword(A<string>._, A<string>._)).Returns(true);
-
-            // When
-            var returnedAdminUser = loginService.GetVerifiedAdminUserAssociatedWithDelegateUsers(
-                delegateUsers,
-                "password"
-            );
-
-            // Then
-            Assert.AreEqual(associatedAdminUser, returnedAdminUser);
-        }
-
-        [Test]
-        public void
-            GetVerifiedAdminUserAssociatedWithDelegateUsers_Returns_verified_admin_account_associated_with_delegate_by_email_with_single_admin_matched_twice()
-        {
-            // Given
-            var delegateUser = UserTestHelper.GetDefaultDelegateUser();
-            var secondDelegateUser = UserTestHelper.GetDefaultDelegateUser(emailAddress: "test@test.com");
-            var delegateUsers = new List<DelegateUser> { delegateUser, delegateUser };
-            var associatedAdminUser = UserTestHelper.GetDefaultAdminUser();
-            A.CallTo(() => userDataService.GetAdminUserByEmailAddress(delegateUser.EmailAddress!))
-                .Returns(associatedAdminUser);
-            A.CallTo(() => userDataService.GetAdminUserByEmailAddress(secondDelegateUser.EmailAddress!))
-                .Returns(associatedAdminUser);
-            A.CallTo(() => cryptoService.VerifyHashedPassword(A<string>._, A<string>._)).Returns(true);
-
-            // When
-            var returnedAdminUser = loginService.GetVerifiedAdminUserAssociatedWithDelegateUsers(
-                delegateUsers,
-                "password"
-            );
-
-            // Then
-            Assert.AreEqual(associatedAdminUser, returnedAdminUser);
-        }
-
-        [Test]
-        public void
-            GetVerifiedAdminUserAssociatedWithDelegateUsers_throws_exception_with_multiple_different_matching_admins()
-        {
-            // Given
-            var delegateUser = UserTestHelper.GetDefaultDelegateUser();
-            var secondDelegateUser = UserTestHelper.GetDefaultDelegateUser(emailAddress: "test@test.com");
-            var delegateUsers = new List<DelegateUser> { delegateUser, secondDelegateUser };
-            var associatedAdminUser = UserTestHelper.GetDefaultAdminUser();
-            var secondAssociatedAdminUser = UserTestHelper.GetDefaultAdminUser(1);
-            A.CallTo(() => userDataService.GetAdminUserByEmailAddress(delegateUser.EmailAddress!))
-                .Returns(associatedAdminUser);
-            A.CallTo(() => userDataService.GetAdminUserByEmailAddress(secondDelegateUser.EmailAddress!))
-                .Returns(secondAssociatedAdminUser);
-            A.CallTo(() => cryptoService.VerifyHashedPassword(A<string>._, A<string>._)).Returns(true);
-
-            // Then
-            Assert.Throws<InvalidOperationException>(
-                () => loginService.GetVerifiedAdminUserAssociatedWithDelegateUsers(
-                    delegateUsers,
-                    "password"
-                )
-            );
-        }
-
-        #endregion
     }
 }
