+++ DigitalLearningSolutions.Data/Services/LoginService.cs
@@ -3,7 +3,6 @@
     using System;
     using System.Collections.Generic;
     using System.Linq;
-    using DigitalLearningSolutions.Data.DataServices.UserDataService;
     using DigitalLearningSolutions.Data.Enums;
     using DigitalLearningSolutions.Data.Models;
     using DigitalLearningSolutions.Data.Models.User;
@@ -11,31 +10,17 @@
     public interface ILoginService
     {
         LoginResult AttemptLogin(string username, string password);
-
-        UserAccountSet VerifyUsers(
-            string password,
-            List<AdminUser> unverifiedAdminUsers,
-            List<DelegateUser> unverifiedDelegateUsers
-        );
-
-        UserAccountSet GetVerifiedLinkedUsersAccounts(
-            int? adminId,
-            int? delegateId,
-            string password
-        );
     }
 
     public class LoginService : ILoginService
     {
-        private readonly ICryptoService cryptoService;
-        private readonly IUserDataService userDataService;
         private readonly IUserService userService;
+        private readonly IUserVerificationService userVerificationService;
 
-        public LoginService(IUserService userService, IUserDataService userDataService, ICryptoService cryptoService)
+        public LoginService(IUserService userService, IUserVerificationService userVerificationService)
         {
             this.userService = userService;
-            this.userDataService = userDataService;
-            this.cryptoService = cryptoService;
+            this.userVerificationService = userVerificationService;
         }
 
         public LoginResult AttemptLogin(string username, string password)
@@ -47,7 +32,7 @@
                 return new LoginResult(LoginAttemptResult.InvalidUsername);
             }
 
-            var (verifiedAdminUsers, verifiedDelegateUsers) = VerifyUsers(
+            var (verifiedAdminUsers, verifiedDelegateUsers) = userVerificationService.VerifyUsers(
                 password,
                 unverifiedAdminUsers,
                 unverifiedDelegateUsers
@@ -111,8 +96,13 @@
 
             var delegateWithEmail = approvedVerifiedDelegates.FirstOrDefault(du => du.EmailAddress != null);
 
-            var verifiedLinkedAdmin = verifiedAdminUser ?? GetVerifiedAdminUserByEmail(password, delegateWithEmail);
-            var verifiedLinkedDelegates = GetVerifiedDelegateUsersAssociatedWithAdminUser(verifiedAdminUser, password);
+            var verifiedLinkedAdmin = verifiedAdminUser ??
+                                      userVerificationService.GetVerifiedAdminUserAssociatedWithDelegateUser(
+                                          delegateWithEmail,
+                                          password
+                                      );
+            var verifiedLinkedDelegates =
+                userVerificationService.GetVerifiedDelegateUsersAssociatedWithAdminUser(verifiedAdminUser, password);
 
             var adminUserToLoginIfCentreActive = new[] { verifiedAdminUser, verifiedLinkedAdmin }
                 .Where(au => au != null).Distinct().SingleOrDefault();
@@ -150,75 +140,6 @@
             }
         }
 
-        public UserAccountSet VerifyUsers(
-            string password,
-            List<AdminUser> unverifiedAdminUsers,
-            List<DelegateUser> unverifiedDelegateUsers
-        )
-        {
-            var verifiedAdminUsers =
-                unverifiedAdminUsers.Where(au => cryptoService.VerifyHashedPassword(au.Password, password))
-                    .ToList();
-            var verifiedDelegateUsers =
-                unverifiedDelegateUsers.Where(du => cryptoService.VerifyHashedPassword(du.Password, password))
-                    .ToList();
-
-            return new UserAccountSet(verifiedAdminUsers, verifiedDelegateUsers);
-        }
-
-        public UserAccountSet GetVerifiedLinkedUsersAccounts(
-            int? adminId,
-            int? delegateId,
-            string password
-        )
-        {
-            var (loggedInAdminUser, loggedInDelegateUser) = userService.GetUsersById(adminId, delegateId);
-
-            var signedInEmailIfAny = loggedInAdminUser?.EmailAddress ?? loggedInDelegateUser?.EmailAddress;
-
-            if (string.IsNullOrWhiteSpace(signedInEmailIfAny))
-            {
-                var loggedInAdminUsers = loggedInAdminUser != null
-                    ? new List<AdminUser> { loggedInAdminUser }
-                    : new List<AdminUser>();
-
-                var loggedInDelegateUsers = loggedInDelegateUser != null
-                    ? new List<DelegateUser> { loggedInDelegateUser }
-                    : new List<DelegateUser>();
-
-                return VerifyUsers(password, loggedInAdminUsers, loggedInDelegateUsers);
-            }
-
-            var (adminUser, delegateUsers) = userService.GetUsersByEmailAddress(signedInEmailIfAny);
-
-            var adminUsers = adminUser != null
-                ? new List<AdminUser> { adminUser }
-                : new List<AdminUser>();
-
-            var temp = VerifyUsers(password, adminUsers, delegateUsers);
-
-            return temp;
-        }
-
-        public AdminUser? GetVerifiedAdminUserAssociatedWithDelegateUsers(
-            List<DelegateUser> delegateUsers,
-            string password
-        )
-        {
-            var adminUsers = new List<AdminUser?>();
-
-            foreach (var delegateUser in delegateUsers)
-            {
-                var admin = GetVerifiedAdminUserByEmail(password, delegateUser);
-                if (admin != null)
-                {
-                    adminUsers.Add(admin);
-                }
-            }
-
-            return adminUsers.Distinct().SingleOrDefault();
-        }
-
         private bool EmailIsTheSameOnAllAccounts(AdminUser? adminUser, List<DelegateUser> delegateUsers)
         {
             if (adminUser == null && delegateUsers.Count == 0)
@@ -238,47 +159,9 @@
             return uniqueEmails.Count == 1;
         }
 
-        private List<DelegateUser> GetVerifiedDelegateUsersAssociatedWithAdminUser(
-            AdminUser? adminUser,
-            string password
-        )
-        {
-            if (string.IsNullOrEmpty(adminUser?.EmailAddress))
-            {
-                return new List<DelegateUser>();
-            }
-
-            var delegatesAssociatedWithAdmin = userDataService.GetDelegateUsersByEmailAddress(adminUser.EmailAddress!);
-
-            var suitableDelegates = delegatesAssociatedWithAdmin
-                .Where(du => du.Active)
-                .Where(du => du.Approved)
-                .Where(du => du.CentreId == adminUser.CentreId)
-                .Where(du => cryptoService.VerifyHashedPassword(du.Password, password));
-
-            return suitableDelegates.ToList();
-        }
-
         private bool NoAccounts(List<AdminUser> adminUsers, List<DelegateUser> delegateUsers)
         {
             return adminUsers.Count == 0 && delegateUsers.Count == 0;
-        }
-
-        private AdminUser? GetVerifiedAdminUserByEmail(string password, DelegateUser? delegateUser)
-        {
-            if (string.IsNullOrWhiteSpace(delegateUser?.EmailAddress))
-            {
-                return null;
-            }
-
-            var adminUserAssociatedWithDelegate = userDataService.GetAdminUserByEmailAddress(delegateUser.EmailAddress);
-
-            var isSuitableAdmin = adminUserAssociatedWithDelegate?.CentreId == delegateUser.CentreId &&
-                                  cryptoService.VerifyHashedPassword(
-                                      adminUserAssociatedWithDelegate.Password,
-                                      password
-                                  );
-            return isSuitableAdmin ? adminUserAssociatedWithDelegate : null;
         }
     }
 }
